Move class `heterogeneous` into directory `memory` from `kernel`
